+++ src/features/main/components/FeaturedArtists.tsx
